+++ TrFunctions.h
@@ -4,7 +4,7 @@ int PaperStoneSissors();
 int add(int num1, int num2, int num3);
 int subtract(int num1, int num2);
 int multiplication(int num1, int num2);
-float division(float num1, float num2);
+float division(int num1, int num2);
 int squareroot();
 int Expo(int base, int expo);
 int factorial(int number);
